feat: alterar tipo de moment_id e tag_id para BIGINT na migração de moment_tags

+++ src/database/migrations/20240205213004-create-moment_tags.js
@@ -11,7 +11,7 @@ module.exports = {
         allowNull: false,
       },
       moment_id: {
-        type: Sequelize.INTEGER(),
+        type: Sequelize.BIGINT(),
         allowNull: false,
         references: {
           model: 'moments', // Certifique-se de ajustar para o nome real da tabela de moments
@@ -21,7 +21,7 @@ module.exports = {
         onDelete: 'CASCADE',
       },
       tag_id: {
-        type: Sequelize.INTEGER(),
+        type: Sequelize.BIGINT(),
         allowNull: false,
         references: {
           model: 'tags', // Certifique-se de ajustar para o nome real da tabela de tags
